refactor: 파일 구조 개선을 위한 파일 경로 변경
